requirements.txt
python-dotenv==1.2.2
schedule==1.2.2
